Move docker image type and correspond util method to plugin lib
