+++ azure-maven-plugin-lib/src/main/java/com/microsoft/azure/maven/appservice/DockerImageType.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.azure.maven.webapp.configuration;
+package com.microsoft.azure.maven.appservice;
 
 public enum DockerImageType {
     NONE,
